pyproject.toml
[tool.poetry]
name = "big-julius-data-ingestion"
version = "1.0.0"
description = "Microservice for ingesting financial transaction CSVs from S3 into Firestore."
authors = ["Big Julius Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
boto3 = "^1.34.0"
google-cloud-firestore = "^2.16.0"
google-auth = "^2.30.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
moto = {extras = ["s3"], version = "^5.0.0"}
ruff = "^0.4.0"
black = "^24.4.0"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=80"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
